fix: move dto directory
