+++ src/main/java/com/driveu/server/domain/question/api/QuestionApi.java
@@ -1,9 +1,9 @@
 package com.driveu.server.domain.question.api;
 
 import com.driveu.server.domain.question.application.QuestionService;
-import com.driveu.server.domain.question.dao.request.QuestionCreateRequest;
-import com.driveu.server.domain.question.dao.response.QuestionListResponse;
-import com.driveu.server.domain.question.dao.response.QuestionResponse;
+import com.driveu.server.domain.question.dto.request.QuestionCreateRequest;
+import com.driveu.server.domain.question.dto.response.QuestionListResponse;
+import com.driveu.server.domain.question.dto.response.QuestionResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
